+++ src/ui/menus/DialogManager.js
@@ -339,7 +339,6 @@ export async function showGridSampleDialog(soundObj, midiNote) {
 			}
 		},
 		onCommit: () => {
-			delete soundObj._eligibleGridKeys;
 			AppState.dispatch({ type: 'AUDIO_UPDATE_REQUESTED' });
 		}
 	});
@@ -409,7 +408,6 @@ export async function showGridSampleDialog(soundObj, midiNote) {
 					speedMax: speedValues.high
 				};
 			}
-			delete soundObj._eligibleGridKeys;
 			AppState.dispatch({ type: 'AUDIO_UPDATE_REQUESTED' });
 			cleanup();
 			resolve(false);